todas as rotas garantem que o usuário está autentc

+++ server/src/routes/memories.ts
@@ -1,11 +1,12 @@
-import { FastifyInstance } from "fastify"
-import { z } from 'zod'
-import { prisma } from '../lib/prisma'
+import { FastifyInstance } from "fastify";
+import { z } from 'zod';
+import { prisma } from '../lib/prisma';
+
 
 export async function memoriesRoutes(app: FastifyInstance) {
   app.addHook('preHandler', async (request) => {
-    await request.jwtVerify()
-  })
+    await request.jwtVerify();
+  });
 
   app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
@@ -15,31 +16,35 @@ export async function memoriesRoutes(app: FastifyInstance) {
       orderBy: {
         createdAt: 'asc',
       },
-    })
-  
+    });
+
     return memories.map((memory) => {
       return {
         id: memory.id,
         coverUrl: memory.coverUrl,
         excerpt: memory.content.substring(0, 115).concat('...'),
-      }
-    })
-  })
+      };
+    });
+  });
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    })
+    });
 
-    const { id } = paramsSchema.parse(request.params)
+    const { id } = paramsSchema.parse(request.params);
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
-    })
+    });
 
-    return memory
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    return memory;
   })
 
   app.post('/memories', async (request) => {
@@ -47,9 +52,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
-    })
+    });
 
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
     const memory = await prisma.memory.create({
       data: {
@@ -58,25 +63,35 @@ export async function memoriesRoutes(app: FastifyInstance) {
         isPublic,
         userId: request.user.sub,
       },
-    })
+    });
 
-    return memory
-  })
+    return memory;
+  });
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    })
+    });
 
-    const { id } = paramsSchema.parse(request.params)
+    const { id } = paramsSchema.parse(request.params);
 
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
+    });
+
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
     })
 
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.update({
       where: {
@@ -87,21 +102,31 @@ export async function memoriesRoutes(app: FastifyInstance) {
         coverUrl,
         isPublic,
       },
-    })
-  })
+    });
+  });
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    })
+    });
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
         id,
       }
-    })
-  })
+    });
+  });
 
 }
